requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-filter==24.3
Pillow==10.4.0